feat: added bru.sleep shim (#2858)
* fix: added console fn for cli

* wip: added bru.sleep shim

---------
